Markers, Routing Stability
Elements that are closer to the desired distance are shown first. If the user doesn't want to go on that route, they can re-roll for a different route (that is the next closest to their desired distance). If they go through every route that we've found, it loops back to the start. The UI is terrible, but the routing is basically done at this point.