BUG-FIX: Rendering a route, loading icon
The loading icon was pushing content to the right, so I shortened "meters" to just "m" and reduced the margin-right of the loading icon. Also, I removed a line from App.css that was causing rendering issues in Challenge.jsx. Since the margin was affecting all images, the map wasn't lining up with the polyline.